Ensure size parameter for GROW_ALLOCATION macro is an int

+++ utils.h
@@ -35,7 +35,10 @@ int errno_would_block(void);
 
 const char *get_errno_name(int error_code);
 
-#define GROW_ALLOCATION(size) ((((size) - 1) / 16 + 1) * 16)
+// round SIZE up to the next multiple of 16. the calculation relies on SIZE
+// being a signed int. with float the division would not truncate the result.
+// with unsigned int SIZE - 1 would overflow to a big value if size is 0.
+#define GROW_ALLOCATION(size) ((((int)(size) - 1) / 16 + 1) * 16)
 
 void string_copy(char *destination, const char *source, int destination_length);
 void string_append(char *destination, const char *source, int destination_length);
